+++ lib/credentials/attestOffchain.ts
